Move Initaiate into separate dir
the openwq initiate routines have been moved to their own folder and separated.
